Intel allows direct downloading now?
Fixes #10

Intel seems to have changed their download page. They don't have an extra page any more where they require a session cookie to be able to download a file, this means that we can directly download the files now. This makes the purpose of this tool obsolete, because before you needed that cookie to download any file and this couldn't be done with wget or similar tools.

I'll keep this tool and will update it if something does not work
any more, such that we have an uniform way to download intel sde.
(In case the URL or something changes again.)